+++ src/Models/Traits/AssetModelTrait.php
@@ -9,6 +9,7 @@ use Hirtz\Media\Models\Asset;
 use Hirtz\Media\Models\Interfaces\AssetModelInterface;
 use Hirtz\Media\Models\Queries\AssetQuery;
 use Hirtz\Skeleton\Models\Traits\TypeAttributeTrait;
+use yii\helpers\Inflector;
 
 /**
  * @mixin AssetModelInterface
@@ -67,5 +68,10 @@ trait AssetModelTrait
     public function getAssetTransformationNames(): array
     {
         return $this->getTypeOptions()['transformations'] ?? [];
+    }
+
+    public function getParamName(): string
+    {
+        return Inflector::slug($this->formName());
     }
 }