+++ src/Modules/Admin/Widgets/Grids/AssetGridView.php
@@ -147,6 +147,11 @@ class AssetGridView extends GridView
      */
     protected function can(string $action, Asset $asset): bool
     {
-        return $this->webuser->can($asset->getPermissionName($action), ['asset' => $asset]);
+        $model = $asset->model;
+
+        return $this->webuser->can($asset->getPermissionName($action), [
+            'asset' => $asset,
+            $model->getParamName() => $model,
+        ]);
     }
 }